Split vcf stuff into multi-file submodule
