+++ chord_variant_service/tables/routes.py
@@ -6,7 +6,7 @@ from flask import Blueprint, current_app, json, jsonify, request, url_for
 from jsonschema import validate, ValidationError
 
 from chord_variant_service.tables.base import TableManager
-from chord_variant_service.tables.vcf import VCFVariantTable
+from chord_variant_service.tables.vcf.table import VCFVariantTable
 from chord_variant_service.tables.exceptions import IDGenerationFailure
 from chord_variant_service.variants.schemas import VARIANT_SCHEMA, VARIANT_TABLE_METADATA_SCHEMA
